Commit Add LoadingView and Nav Link to Item Menu

+++ DirtyApron/Supporting Views/AdminViews/MenuItemView.swift
@@ -15,56 +15,69 @@ struct MenuItemView: View {
     @State var menuItem = MenuItem()
     @State private var showingAddMenuItem = false
     @State private var isEdit = false
+    @State private var showingAlert = false
+    @State private var loading = false
+    @State private var message = ""
     @State var pickedType = [String]()
     var category: Category!
     
     var body: some View {
-        List {
-            ForEach(menuItems.lists, id: \.id) { menuItem in
-                NavigationLink(destination: DetailItemView(menuItem: menuItem)){
-                    HStack {
-                        VStack(alignment: .leading, spacing: 5) {
-                            Text(menuItem.name)
-                                .fontWeight(menuItem.isEnable ? .bold : .none)
-                                .foregroundColor(menuItem.isEnable ? .primary : .secondary)
-                                .font(.headline)
-                                .bold()
+        ZStack {
+            if menuItems.lists.isEmpty && loading {
+                withAnimation {
+                    LoadingView(text: "Loading...")
+                }
+                .animation(.easeOut)
+            } else {
+                List {
+                    ForEach(menuItems.lists, id: \.id) { menuItem in
+                        NavigationLink(destination: DetailItemView(menuItem: menuItem)){
                             HStack {
-                                ForEach(menuItem.foodType, id: \.self) {
-                                    Text($0)
-                                        .badgesStyle(text: $0)
+                                VStack(alignment: .leading, spacing: 5) {
+                                    Text(menuItem.name)
+                                        .fontWeight(menuItem.isEnable ? .bold : .none)
+                                        .foregroundColor(menuItem.isEnable ? .primary : .secondary)
+                                        .font(.headline)
+                                        .bold()
+                                    HStack {
+                                        ForEach(menuItem.foodType.sorted(), id: \.self) {
+                                            Text($0)
+                                                .badgesStyle(text: $0)
+                                        }
+                                    }
+                                }
+                                
+                                Spacer()
+                                Group {
+                                    if menuItem.amount != 0 {
+                                        //                                Text("£\(menuItem.amount, specifier: "%.2f")")
+                                        Button(action : {
+                                            print("\(menuItem.name) - £\(menuItem.amount)")
+                                        }) {
+                                            
+                                            Text("£\(menuItem.amount, specifier: "%.2f")")
+                                                .styleButton(colour: .blue)
+                                            
+                                        }
+                                        .padding(.leading)
+                                        .buttonStyle(PlainButtonStyle())
+                                    } else {
+                                        Text(" INFO")
+                                    }
                                 }
                             }
-                        }
-                        
-                        Spacer()
-                        Group {
-                            if menuItem.amount != 0 {
-//                                Text("£\(menuItem.amount, specifier: "%.2f")")
-                                Button(action : {
-                                    print("\(menuItem.name) - £\(menuItem.amount)")
-                                }) {
-                                    
-                                    Text("£\(menuItem.amount, specifier: "%.2f")")
-                                        .styleButton(colour: .blue)
-                                    
-                                }
-                                .padding(.leading)
-                                .buttonStyle(PlainButtonStyle())
-                            } else {
-                                Text(" INFO")
+                            .onTapGesture(count: 2) {
+                                self.menuItem = menuItem
+                                self.isEdit = true
+                                self.showingAddMenuItem.toggle()
                             }
                         }
                     }
-                    .onTapGesture(count: 2) {
-                        self.menuItem = menuItem
-                        self.isEdit = true
-                        self.showingAddMenuItem.toggle()
-                    }
+                    .onDelete(perform: deleteItem)
                 }
             }
-            .onDelete(perform: deleteItem)
         }
+        
         .onAppear(perform: loadItems)
         .sheet(isPresented: $showingAddMenuItem) {
             AddMenuItemView(menuItems: self.menuItems, menuItem: self.menuItem, amount: (self.menuItem.amount == 0 ? "" : String(self.menuItem.amount)), category: self.category, isEdit: self.isEdit)
@@ -85,15 +98,18 @@ struct MenuItemView: View {
     }
     
     //MARK: Fetch Menu Items
-    func loadItems() {
+    private func loadItems() {
+        loading.toggle()
         if let recordID = category.recordID {
             CKHelper.fetchItems(recordID: recordID) { (results) in
                 switch results {
                 case .success(let newItems):
                     self.menuItems.lists = newItems
                 case .failure(let error):
-                    print(error.localizedDescription)
+                    self.message = error.localizedDescription
+                    self.showingAlert.toggle()
                 }
+                self.loading.toggle()
             }
         }
     }
@@ -106,7 +122,8 @@ struct MenuItemView: View {
         CKContainer.default().publicCloudDatabase.delete(withRecordID: recordID) { (recordID, error) in
             DispatchQueue.main.async {
                 if let error = error {
-                    print(error.localizedDescription)
+                    self.message = error.localizedDescription
+                    self.showingAlert.toggle()
                 } else {
                     self.menuItems.lists.remove(at: index)
                 }
